deleting files now also deletes the ratings

+++ list/delete.php
@@ -9,18 +9,22 @@
     $userId = $_SESSION["userId"];
     
     if($userId==0){
-        $sql = $conn->prepare("DELETE FROM files WHERE name = ?");
+        $sql = $conn->prepare("DELETE FROM files WHERE files.name = ?");
         $sql->bind_param('s', $id);
+
     }
     else{
-        $sql = $conn->prepare("DELETE FROM files WHERE name = ? AND userId = ?");
+        $sql = $conn->prepare("DELETE FROM userrating,files WHERE files.name = ? AND userId = ?");
         $sql->bind_param("si", $id, $userId);
     }
 
     $sql->execute();
-    $conn->close();
     if($sql->affected_rows>0){//if rows got deleted
+        $sql = $conn->prepare("DELETE FROM userrating WHERE fileId = ?");
+        $sql->bind_param('s', $id);
+        $sql->execute();
         unlink('../thumbnails/'.$id);
         unlink('../files/'.$id);
     }
+    $conn->close();
 ?>
